4/1 Changes (2) Found that the color mixing being used on the site was more complex then expected * rgb_to_cymk() created to acount for the more complex color mixing * cymk_to_rgb() created to acount for the more complex color mixing * path_color() takes all the tiles in a path and returns their mixed color * mix_colors() has been changed to acount for complex color mixing
find_path can now find all possible paths to the end tile from the start tile.
It will stop itself from finding paths that are longer then the max travel
The problem with this is that the later levels could have a large number
of possible paths, I need to get the color mixing to work.